requirements.txt
aiohttp==3.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
Automat==0.7.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
Flask==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
hyperlink==18.0.0
idna==2.7
idna-ssl==1.1.0
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.0
multidict==4.3.1
parsel==1.4.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyMySQL==0.8.1
pyOpenSSL==18.0.0
pyquery==1.4.0
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
Scrapy==1.5.0
selenium==3.12.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.10
flask-apscheduler==1.10.1
Twisted==18.4.0
schedule==0.5.0
urllib3==1.23
w3lib==1.19.0
Werkzeug==0.14.1
yarl==1.2.6
zope.interface==4.5.0
flask_sqlalchemy==2.3.2
kafka-python==1.4.6
openpyxl==2.5.11
pandas==0.23.4
numpy==1.15.4
Flask-Redis==0.3.0
xlwt==1.3.0
raven==6.10.0
blinker==1.4
Flask-Log-Request-ID==0.10.0
PyExecJS==1.5.1